backend/requirements.txt
# --- Core Framework & Server ---
Flask
flask-cors

# --- Deep Learning (PyTorch) ---
torch
torchvision
# Provides both MTCNN (detector/aligner) and FaceNet (embedder)
facenet-pytorch 

# --- Image & Video Processing ---
opencv-python-headless
Pillow

# --- Data Handling & Reporting ---
numpy
pandas
fpdf2